- added support for y_inline - fixed some memory leaks - improved Windows XP compatibility - fixed bug where OnQueryError parameters had empty values - renamed mysql_real_escape_string to mysql_escape_string - rewrote native mysql_format - reduced usage of memory allocation (malloc) - added enum datatype support to int-conversion natives - replaced mysql_function_query native wrapper with pawn define - changed MySQL connector again (sorry for that, Windows users will have to update the libmysql.dll again) - improved HTML log file design

+++ trunk/src/mysql_include/my_alloc.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2000 MySQL AB
+/* Copyright (c) 2000, 2002, 2005, 2007 MySQL AB
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -26,8 +26,8 @@
 typedef struct st_used_mem
 {				   /* struct for once_alloc (block) */
   struct st_used_mem *next;	   /* Next block in use */
-  size_t left;                     /* memory left in block  */
-  size_t size;                     /* size of block */
+  unsigned int	left;		   /* memory left in block  */
+  unsigned int	size;		   /* size of block */
 } USED_MEM;
